[room-server] Publish & subscribe endpoints, shape media
Separates the types of WebRtc endpoints into publishers (whose media
could be modified before and during streaming) and subscribers.
*Note: media shaping during streaming has not been implemented yet.

The TrickleICEWebRtcEndpint interface is removed, but its methods
are defined in the IceWebRtcEndpoint abstract class.

Adds 'publishVideo' method to the room manager and the consequent
'publishToRoom' method in the Participant class. 'publishVideo' invokes
a method from the SessionInterceptor, allowing its implementor to shape
the media stream before it is published. This type of request should
be performed when the user decides to publish her local stream into the room.

Change-Id: I32493bdd13717cfabfb955ef0e3caa592a814f51