[room] Unpublish and unsubscribe methods, some fixes
Adds these methods to the SDK manager, with the corresponding events.
Renames receiveMedia to subscribe and adds room's name as parameter
to some events.
Changes room demo to implement the new functionality.
Adds the new Websocket methods and changes the names of some older methods.
Fixes the test app to close the websocket when leaving the room (calls close
from KurentoRoom instead of just leaving the room).

Change-Id: I99c838deee9f3f42e4d435b5cd297a0066498a1c