[room-server,demo] Custom WS request, hat toggle in demo
Adds a custom request in the WebSocket API of the server.
The demo uses the custom request to toggle the hat filter.
Demo configuration of the local video display of the web clients.
Adds another hat (wizard) and changes the demo configuration.
Removes the "first publisher has a hat" option.

Change-Id: Ic6ea02b0b3177377ae21e3e986465f0225bd05b0