[room-test] Initial import of the room test framework
The new module is based on the kurento-java test framework.
Separates the project's version from the kurento-parent-pom one.
Updates kurento-parent-pom version to 6.2.2-SNAPSHOT.
Modifies the tests in kurento-room-demo-test to use the new framework.
Ignores tests in kurento-room-test until the framework gets stabilized.
Adds 'test.app.classname' property to indicate the web server that will
be executed as a service during the tests (default is KurentoRoomDemoApp).

Change-Id: Ic6d592686233e0e81898d5692ab8870d7153615b