[room-demo] Fix back and fw browser buttons, access control
When pressing the Back button, or refreshing the page when in a conference,
it closes the connection (acts as leave from room) and redirects to login.
When trying to access the call page (/#/call) it first checks if the
user is a participant in a room; if not, redirects to login.
All other routes are redirected to login.

Change-Id: I56f4f7c9b1adcb9569083146c9f04c3023478e3b